Refactored to allow testing of items other than fonts
